Update Footer component import in App.jsx and Home.jsx, style added.
