+++ src/App.jsx
@@ -1,7 +1,7 @@
 import Header from "./components/Header/Header"
 import { Route, Routes } from 'react-router-dom'
 import Home from "./components/Home/Home"
-import Footer from "./components/Home/Footer"
+import Footer from "./components/Footer/Footer"
 
 function App() {
